[백엔드 :: 엔티티 PK 타입 Integer -> Long]

+++ src/main/java/com/example/oose_23_3_back/historycontrol/entity/HistoryCheck.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 public class HistoryCheck {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer historyId;
+    private Long historyId;
     @ManyToOne
     private Member member;
     @ManyToOne
